Add number video view

+++ app/src/main/java/com/orange/signsatwork/biz/domain/Video.java
@@ -31,7 +31,7 @@ public class Video {
     public final long id;
     public final String url;
     public final String pictureUri;
-    public final double nbView;
+    public final long nbView;
     public final Date createDate;
     public final User user;
     public final Sign sign;
